Match the reflection engine to the conversion-safety doctrine
The reflection path bounded sizes at the uint32 range structurally: the
descriptor Size/Len/BitSize/MinSize fields were uint32, getSszValueSize
returned uint32, and dynssz-size resolution rejected values above
math.MaxUint32. The generated code allows any size up to the platform
integer range as long as no 4-byte offset overflows, so the engines
disagreed on sizes in between — and reflection marshal wrote silently
truncated offsets past 4 GiB.

- Descriptor Size/Len/BitSize/MinSize (and DynFieldDescriptor's
  HeaderOffset) widen to int64; size products and container sums are
  bounded to the platform integer range with wrap-safe checks.
- dynssz-size resolution and size annotations reject values above
  math.MaxInt with ErrPlatformOverflow, matching the generated guard;
  static ssz-size literals parse into the full range.
- getSszValueSize and ReflectionCtx.SizeSSZ return int64.
- Reflection marshal writes offsets through checked helpers: a value
  past the 4-byte SSZ offset range returns ErrOffsetOverflowFn instead
  of truncating, in the container, vector, and list paths.

Sizes past the former uint32 bound now behave identically in both
engines (differential fuzzing clean); tests asserting the old bound are
updated to the platform-range doctrine, including a regression test that
a 2^32-byte vector type is accepted.

+++ codegen/gen_decoder.go
@@ -520,7 +520,7 @@ func (ctx *decoderContext) unmarshalTypeWrapper(desc *ssztypes.TypeDescriptor, v
 
 // unmarshalContainer generates unmarshal code for SSZ container (struct) types.
 func (ctx *decoderContext) unmarshalContainer(desc *ssztypes.TypeDescriptor, varName string, typePath typePathList, indent int) error {
-	staticSize := uint32(0)
+	staticSize := int64(0)
 	staticSizeVars := []string{}
 	hasDynamicFields := false
 	for _, field := range desc.ContainerDesc.Fields {
